backend/requirements.txt
openai
python-dotenv
chromadb
sentence-transformers
uvicorn
fastapi
numpy
requests
